requirements.txt
requests==2.28.1
jupyterlab==3.4.7
pyspark==3.5.0
PyYAML==6.0
ipykernel==6.15.0
pandas==1.3.5
pymongo[srv]
python-dotenv
apache-airflow
streamlit
pyarrow
-e .

